edamam api (fixed)
prototype with edamam api and functioning search page
