+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.7.3",
     "bootstrap": "^4.6.0",
+    "cors": "^2.8.5",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.3",
@@ -39,5 +40,6 @@
   },
   "devDependencies": {
     "react-router-dom": "^5.2.0"
-  }
+  },
+  "proxy": "https://api.edamam.com"
 }
