--- public/src/index.css
@@ -1,113 +1,0 @@
-@import url(https://fonts.googleapis.com/css2?family=Roboto&display=swap);
-@import url(https://fonts.googleapis.com/css2?family=Rubik:wght@300;400;500;700&display=swap);
-
-
-body {
-  padding-top: 30px;
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-/* Theme Colors */
-:root {
-  --surface: #FAF6EF;
-  --surface-dark: #ECDAC1;
-  --text: #010C14;
-  --text-light: #03233A;
-  --primary: #D07D62;
-  --primary-light: #DEA491;
-  --primary-dark: #AD5234;
-  --secondary: #496F5B;
-  --secondary-light: #A9C7B7;
-  --secondary-dark: #283E32;
-  --accent: #04395E;
-  --danger: #C71A37;
-  --danger-light: #EA5D74;
-}
-
-.container-fluid {
-  background-color: var(--surface);
-}
-
-.h1 {
-  font-family: "Rubik", sans-serif;
-  color: var(--secondary);
-}
-
-.h2 {
-  font-family: "Rubik", sans-serif;
-  color: var(--secondary)
-}
-
-.h3 {
-  font-family: "Rubik", sans-serif;
-  color: var(--text);
-}
-
-.p {
-  font-family: "Roboto", sans-serif;
-  color: var(--text)
-}
-
-.wbdv-link-text {
-  font-family: "Roboto", sans-serif;
-  color: var(--accent);
-  font-style: oblique;
-}
-
-.wbdv-detail-btn {
-  background-color: var(--secondary);
-  font-family: "Roboto", sans-serif;
-  color: var(--surface);
-}
-
-.wbdv-affirmative-btn {
-  background-color: var(--primary);
-  font-family: "Roboto", sans-serif;
-  color: var(--surface);
-}
-
-.wbdv-secondary-btn {
-  background-color: var(--accent);
-  font-family: "Roboto", sans-serif;
-  color: var(--surface);
-}
-
-.wbdv-navigation-bar {
-  padding: 10px;
-  font-weight: bold;
-  font-size: larger;
-  position: fixed;
-  width: 100%;
-  z-index: 1;
-  background-color: var(--primary);
-  color: var(--surface);
-  font-family: "Rubik", sans-serif;
-}
-
-.wbdv-center-in-div {
-  text-align: center;
-  display: flex;
-  justify-content: center;
-  flex-direction: column-reverse;
-}
-
-.container-fluid {
-  padding: 0;
-}
-
-.wbdv-image {
-  /*border-radius: 70%;*/
-  width: 400px;
-  height: 250px;
-  padding: 10px;
-}
