+++ src/App.js
@@ -2,8 +2,11 @@ import logo from './logo.svg';
 import './App.css';
 import {BrowserRouter} from "react-router-dom";
 import BaseApp from "./components/base-app";
+import React from "react";
 
-function App() {
+
+const App = () => {
+
   return (
       <BrowserRouter>
         <BaseApp/>
