+++ src/components/details-page/details-modules.css
