+++ src/components/details-page/details-page.js
